Rename App.{cpp,h} to OrbitApp.{cpp,h}
The contained class is called OrbitApp, so should be the files.
