+++ src/OrbitGl/AsyncTrack.cpp
@@ -19,9 +19,9 @@
 #include "ClientProtos/capture_data.pb.h"
 #include "DisplayFormats/DisplayFormats.h"
 #include "OrbitBase/Logging.h"
-#include "OrbitGl/App.h"
 #include "OrbitGl/GlUtils.h"
 #include "OrbitGl/ManualInstrumentationManager.h"
+#include "OrbitGl/OrbitApp.h"
 #include "OrbitGl/PrimitiveAssembler.h"
 #include "OrbitGl/TextRenderer.h"
 #include "OrbitGl/TimeGraph.h"
