+++ src/OrbitGl/CallstackThreadBar.cpp
@@ -20,12 +20,12 @@
 #include "ClientData/FastRenderingUtils.h"
 #include "OrbitBase/Logging.h"
 #include "OrbitBase/ThreadConstants.h"
-#include "OrbitGl/App.h"
 #include "OrbitGl/BatcherInterface.h"
 #include "OrbitGl/CoreMath.h"
 #include "OrbitGl/FormatCallstackForTooltip.h"
 #include "OrbitGl/Geometry.h"
 #include "OrbitGl/GlCanvas.h"
+#include "OrbitGl/OrbitApp.h"
 #include "OrbitGl/PickingManager.h"
 #include "OrbitGl/PrimitiveAssembler.h"
 #include "OrbitGl/ThreadColor.h"
