+++ src/OrbitGl/CaptureWindow.cpp
@@ -31,13 +31,13 @@
 #include "OrbitBase/Profiling.h"
 #include "OrbitBase/Result.h"
 #include "OrbitBase/ThreadConstants.h"
-#include "OrbitGl/App.h"
 #include "OrbitGl/BatcherInterface.h"
 #include "OrbitGl/CaptureViewElement.h"
 #include "OrbitGl/CoreMath.h"
 #include "OrbitGl/Geometry.h"
 #include "OrbitGl/GlUtils.h"
 #include "OrbitGl/OpenGlBatcher.h"
+#include "OrbitGl/OrbitApp.h"
 #include "OrbitGl/PrimitiveAssembler.h"
 #include "OrbitGl/QtTextRenderer.h"
 #include "OrbitGl/TextRenderer.h"
