+++ src/OrbitGl/GpuDebugMarkerTrack.cpp
@@ -13,8 +13,8 @@
 #include "ClientProtos/capture_data.pb.h"
 #include "DisplayFormats/DisplayFormats.h"
 #include "OrbitBase/Logging.h"
-#include "OrbitGl/App.h"
 #include "OrbitGl/GlUtils.h"
+#include "OrbitGl/OrbitApp.h"
 #include "OrbitGl/PrimitiveAssembler.h"
 #include "OrbitGl/TimeGraph.h"
 #include "OrbitGl/TimeGraphLayout.h"
