+++ src/OrbitGl/SchedulerTrack.cpp
@@ -18,8 +18,8 @@
 #include "ClientProtos/capture_data.pb.h"
 #include "OrbitBase/Logging.h"
 #include "OrbitBase/ThreadConstants.h"
-#include "OrbitGl/App.h"
 #include "OrbitGl/BatcherInterface.h"
+#include "OrbitGl/OrbitApp.h"
 #include "OrbitGl/PrimitiveAssembler.h"
 #include "OrbitGl/ThreadColor.h"
 #include "OrbitGl/TimeGraphLayout.h"
