+++ src/OrbitGl/ThreadStateBar.cpp
@@ -23,7 +23,6 @@
 #include "DisplayFormats/DisplayFormats.h"
 #include "GrpcProtos/capture.pb.h"
 #include "OrbitBase/Logging.h"
-#include "OrbitGl/App.h"
 #include "OrbitGl/AsyncTrack.h"
 #include "OrbitGl/BatcherInterface.h"
 #include "OrbitGl/CaptureViewElement.h"
@@ -32,6 +31,7 @@
 #include "OrbitGl/Geometry.h"
 #include "OrbitGl/GlCanvas.h"
 #include "OrbitGl/GlUtils.h"
+#include "OrbitGl/OrbitApp.h"
 #include "OrbitGl/PrimitiveAssembler.h"
 #include "OrbitGl/ThreadBar.h"
 #include "OrbitGl/Viewport.h"
