+++ src/OrbitGl/ThreadTrack.cpp
@@ -24,11 +24,11 @@
 #include "OrbitBase/Logging.h"
 #include "OrbitBase/ThreadConstants.h"
 #include "OrbitBase/Typedef.h"
-#include "OrbitGl/App.h"
 #include "OrbitGl/BatcherInterface.h"
 #include "OrbitGl/Geometry.h"
 #include "OrbitGl/GlCanvas.h"
 #include "OrbitGl/GlUtils.h"
+#include "OrbitGl/OrbitApp.h"
 #include "OrbitGl/PrimitiveAssembler.h"
 #include "OrbitGl/TextRenderer.h"
 #include "OrbitGl/ThreadColor.h"
