+++ src/OrbitGl/TimeGraph.cpp
@@ -23,7 +23,6 @@
 #include "OrbitBase/Logging.h"
 #include "OrbitBase/Typedef.h"
 #include "OrbitGl/AccessibleTimeGraph.h"
-#include "OrbitGl/App.h"
 #include "OrbitGl/AsyncTrack.h"
 #include "OrbitGl/CGroupAndProcessMemoryTrack.h"
 #include "OrbitGl/FrameTrack.h"
@@ -31,6 +30,7 @@
 #include "OrbitGl/GlCanvas.h"
 #include "OrbitGl/GlUtils.h"
 #include "OrbitGl/GpuTrack.h"
+#include "OrbitGl/OrbitApp.h"
 #include "OrbitGl/PageFaultsTrack.h"
 #include "OrbitGl/PickingManager.h"
 #include "OrbitGl/SchedulerTrack.h"
