+++ src/OrbitGl/TimerTrack.cpp
@@ -23,9 +23,9 @@
 #include "DisplayFormats/DisplayFormats.h"
 #include "OrbitBase/Logging.h"
 #include "OrbitBase/Typedef.h"
-#include "OrbitGl/App.h"
 #include "OrbitGl/Geometry.h"
 #include "OrbitGl/GlCanvas.h"
+#include "OrbitGl/OrbitApp.h"
 #include "OrbitGl/PrimitiveAssembler.h"
 #include "OrbitGl/TextRendererInterface.h"
 #include "OrbitGl/TimeGraphLayout.h"
