+++ src/OrbitGl/TracepointThreadBar.cpp
@@ -16,12 +16,12 @@
 #include "ClientData/TracepointInfo.h"
 #include "OrbitBase/Logging.h"
 #include "OrbitBase/ThreadConstants.h"
-#include "OrbitGl/App.h"
 #include "OrbitGl/BatcherInterface.h"
 #include "OrbitGl/CaptureViewElement.h"
 #include "OrbitGl/CoreMath.h"
 #include "OrbitGl/Geometry.h"
 #include "OrbitGl/GlCanvas.h"
+#include "OrbitGl/OrbitApp.h"
 #include "OrbitGl/PrimitiveAssembler.h"
 #include "OrbitGl/ThreadColor.h"
 #include "OrbitGl/TimeGraphLayout.h"
