+++ src/OrbitGl/TrackContainer.cpp
@@ -28,12 +28,12 @@
 #include "OrbitBase/Logging.h"
 #include "OrbitBase/Sort.h"
 #include "OrbitGl/AccessibleCaptureViewElement.h"
-#include "OrbitGl/App.h"
 #include "OrbitGl/BatcherInterface.h"
 #include "OrbitGl/CoreMath.h"
 #include "OrbitGl/Geometry.h"
 #include "OrbitGl/GlCanvas.h"
 #include "OrbitGl/GlUtils.h"
+#include "OrbitGl/OrbitApp.h"
 #include "OrbitGl/PickingManager.h"
 #include "OrbitGl/ThreadColor.h"
 #include "OrbitGl/ThreadTrack.h"
