+++ src/OrbitGl/TrackManager.cpp
@@ -31,8 +31,8 @@
 #include "OrbitBase/Logging.h"
 #include "OrbitBase/Sort.h"
 #include "OrbitBase/ThreadConstants.h"
-#include "OrbitGl/App.h"
 #include "OrbitGl/CoreMath.h"
+#include "OrbitGl/OrbitApp.h"
 #include "OrbitGl/TimeGraphLayout.h"
 #include "OrbitGl/TrackContainer.h"
 #include "OrbitGl/Viewport.h"
