+++ src/OrbitQt/orbitmainwindow.cpp
@@ -106,12 +106,12 @@
 #include "OrbitBase/Result.h"
 #include "OrbitBase/StopToken.h"
 #include "OrbitGgp/Instance.h"
-#include "OrbitGl/App.h"
 #include "OrbitGl/CaptureWindow.h"
 #include "OrbitGl/DataViewFactory.h"
 #include "OrbitGl/GlCanvas.h"
 #include "OrbitGl/IntrospectionWindow.h"
 #include "OrbitGl/LiveFunctionsController.h"
+#include "OrbitGl/OrbitApp.h"
 #include "OrbitGl/SamplingReport.h"
 #include "OrbitGl/TimeGraph.h"
 #include "OrbitGl/TrackManager.h"
