+++ src/OrbitQt/orbitmainwindow.h
@@ -53,9 +53,9 @@
 #include "OrbitBase/MainThreadExecutor.h"
 #include "OrbitBase/Result.h"
 #include "OrbitBase/StopToken.h"
-#include "OrbitGl/App.h"
 #include "OrbitGl/CallTreeView.h"
 #include "OrbitGl/MainWindowInterface.h"
+#include "OrbitGl/OrbitApp.h"
 #include "OrbitGl/StaticTimeGraphLayout.h"
 #include "SessionSetup/TargetConfiguration.h"
 #include "SessionSetup/TargetLabel.h"
